requirements.txt
django
djangorestframework
pillow
psycopg2-binary
python-decouple
pyjwt
drf-yasg
django-filter
djangorestframework-simplejwt
django-rest-swagger